refactor(@dpc-sdp/ripple-tide-news): :recycle: refactor news content type to use nuxt layers
