+++ packages/ripple-tide-news/components/TideNewsHeader.vue
@@ -3,19 +3,13 @@ export default { name: 'TideNewsHeader' }
 </script>
 
 <template>
-  <RplHeroHeader
-    :corner-top="true"
-    :corner-bottom="true"
-    :behind-nav="true"
-    :breadcrumbs="hasBreadcrumbs"
-    :title="header.title"
-  >
+  <RplHeaderHeroHeader :corner-top="true" :corner-bottom="true" :behind-nav="true" :breadcrumbs="hasBreadcrumbs"
+    :title="header.title">
     <p class="rpl-type-p-large">{{ header.summary }}</p>
-  </RplHeroHeader>
+  </RplHeaderHeroHeader>
 </template>
 
 <script setup lang="ts">
-import { RplHeroHeader } from '@dpc-sdp/ripple-ui-core'
 import type { TideNewsHeader } from '../types'
 
 defineProps<{
