+++ packages/ripple-tide-news/components/global/TideNews.vue
@@ -1,16 +1,6 @@
-<script lang="ts">
-export default { name: 'TideNewsPage' }
-</script>
-
 <template>
-  <TideBaseLayout
-    :site="site"
-    :pageTitle="page.title"
-    :pageDescription="page.description"
-    :pageLanguage="page.lang"
-    :topicTags="page.showTopicTags ? page.topicTags : []"
-    :updatedDate="page.changed || page.created"
-  >
+  <TideBaseLayout :site="site" :pageTitle="page.title" :pageDescription="page.description" :pageLanguage="page.lang"
+    :topicTags="page.showTopicTags ? page.topicTags : []" :updatedDate="page.changed || page.created">
     <template #aboveHeader>
       <slot name="aboveHeader"></slot>
     </template>
@@ -24,16 +14,9 @@ export default { name: 'TideNewsPage' }
       <TideNewsHeader :header="page.header" :hasBreadcrumbs="hasBreadcrumbs" />
     </template>
     <template #body="{ hasSidebar }">
-      <TideNewsBody
-        :body="page.body"
-        :details="page.details"
-        :components="page.dynamicComponents"
-      />
-      <TideDynamicComponents
-        v-if="page.dynamicComponents?.length > 0"
-        :components="page.dynamicComponents"
-        :hasSidebar="hasSidebar"
-      />
+      <TideNewsBody :body="page.body" :details="page.details" :components="page.dynamicComponents" />
+      <TideDynamicComponents v-if="page.dynamicComponents?.length > 0" :components="page.dynamicComponents"
+        :hasSidebar="hasSidebar" />
     </template>
     <template #sidebar>
       <slot name="sidebar"></slot>
@@ -46,9 +29,7 @@ export default { name: 'TideNewsPage' }
 
 <script setup lang="ts">
 import { TideSiteData } from '@dpc-sdp/ripple-tide-api/types'
-import type { TideNewsPage } from './types'
-import TideNewsHeader from './components/tide-news-header.vue'
-import TideNewsBody from './components/tide-news-body.vue'
+import type { TideNewsPage } from '../../types'
 
 interface Props {
   site: TideSiteData
