+++ packages/ripple-tide-news/nuxt.config.ts
@@ -1,0 +1,3 @@
+import { defineNuxtConfig } from 'nuxt/config'
+
+export default defineNuxtConfig({})
