+++ packages/ripple-tide-news/server/middleware/tideNewsMiddleware.ts
@@ -1,0 +1,7 @@
+import { defineEventHandler } from 'h3'
+import { registerTideContentType } from '@dpc-sdp/ripple-tide-api'
+import tideNewsModule from '../../mapping'
+
+export default defineEventHandler(async (event) => {
+  registerTideContentType(event, 'news', tideNewsModule)
+})
